requirements.txt
# Core orchestration
langgraph>=0.0.46

# LLM + API integration
openai>=1.40.0
requests>=2.31.0

# Document processing (Stage 2 normalization)
docx2python>=2.0.0
mammoth>=1.6.0
pdfminer.six>=20221105
unstructured[local-inference]>=0.10.30
textract-py3>=1.6.4   # replaces 'textract', avoids metadata issues

# XLSX / DOCX outputs
pandas>=2.2.2
openpyxl>=3.1.3
python-docx>=1.0.1

# Schema validation
pydantic>=2.8.2

# Text/vector similarity (deduplication)
scikit-learn>=1.5.0
numpy>=1.26.4

# Env management
python-dotenv>=1.0.1

# Testing (optional)
pytest>=8.2.0
